Move channels to new synchronization module
